requirements.txt
pymupdf 
sentence-transformers 
faiss-cpu 
fastapi 
uvicorn
transformers
